requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
joblib
fastapi
langchain
langchain-google-genai
python-dotenv
pydantic
langchain-huggingface
fastapi
uvicorn
pandas
python-multipart
PyJWT
bcrypt
pydantic[email]